Remove isInFilterContext and workarounds
`this.isInFilterContext`appears to be unused in bodybuilder.js - it gets passed as "this", but apparently there's a scoping issue where it never gets set.

This makes the split of filterBuilder/queryBuilder not make sense. And since there are tests that rely on that behavior, remove the extra classes that workaround this and just pass bodybuilder.